Dont' allow last admin deletion

+++ pkg/query-service/dao/sqlite/rbac.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/pkg/errors"
+	"go.signoz.io/query-service/constants"
 	"go.signoz.io/query-service/model"
 	"go.signoz.io/query-service/telemetry"
 	"go.uber.org/zap"
@@ -271,6 +272,23 @@ func (mds *ModelDaoSqlite) UpdateUserPassword(ctx context.Context, passwordHash,
 func (mds *ModelDaoSqlite) DeleteUser(ctx context.Context, id string) *model.ApiError {
 	zap.S().Debugf("Deleting user. Id: %s\n", id)
 
+	userGroup, apiErr := mds.GetUserGroup(ctx, id)
+	if apiErr != nil {
+		return apiErr
+	}
+	adminGroup, apiErr := mds.GetGroupByName(ctx, constants.AdminGroup)
+	if apiErr != nil {
+		return apiErr
+	}
+	adminUsers, apiErr := mds.GetGroupUsers(ctx, adminGroup.Id)
+	if apiErr != nil {
+		return apiErr
+	}
+
+	if userGroup.GroupId == adminGroup.Id && len(adminUsers) == 1 {
+		return &model.ApiError{Typ: model.ErrorInternal, Err: fmt.Errorf("cannot delete the last admin")}
+	}
+
 	result, err := mds.db.ExecContext(ctx, `DELETE from users where id=?;`, id)
 	if err != nil {
 		zap.S().Errorf("Error while deleting user from the DB\n", err)
@@ -282,7 +300,7 @@ func (mds *ModelDaoSqlite) DeleteUser(ctx context.Context, id string) *model.Api
 		return &model.ApiError{Typ: model.ErrorExec, Err: err}
 	}
 	if affectedRows == 0 {
-		return &model.ApiError{Typ: model.ErrorNotFound, Err: fmt.Errorf("no dashboard found with id: %s", id)}
+		return &model.ApiError{Typ: model.ErrorNotFound, Err: fmt.Errorf("no user found with id: %s", id)}
 	}
 
 	return nil
